pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "verified-destinations"
version = "0.1.0"
description = "Package to house dat verified destinations"
authors = ["datLabs <team@datlabs.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dat-core = {git = "https://github.com/dat-labs/dat-core", rev = "main", develop = true}
# dat-core = {path = "../dat-core", develop = true}
pinecone-client = "^3.1.0"
qdrant-client = "^1.8.0"
weaviate-client = "3.26.2"

[tool.poetry.group.dev.dependencies]
pydantic = "^2.6.3"
pyyaml = "^6.0.1" 
mypy = "^1.8.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
